renaming files, namespace and variable
