+++ Src/Plugins/Tests.Security/SEC1001.SecurePlaintextSecretsTests.cs
