+++ Src/Plugins/Tests.Security/SEC1002.ReviewPotentiallySensitiveDataTests.cs
