+++ Src/Plugins/Tests.Security/SEC1003.ReviewPotentiallySensitiveFilesTests.cs
